Fix build: remove ngrokAuthToken reference, fix offset import

+++ app/src/main/java/com/remoteadb/app/ui/screens/OnboardingScreen.kt
@@ -14,7 +14,6 @@ import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.draw.offset
 import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
@@ -168,7 +167,7 @@ private fun OnboardingPageContent(page: OnboardingPage) {
         Box(
             modifier = Modifier
                 .size(160.dp)
-                .offset(y = offsetY.dp),
+                .offset(y = offsetY.toInt().dp),
             contentAlignment = Alignment.Center
         ) {
             Box(
